\name{scCNA-package}
\alias{scCNA-package}
\alias{scCNA}
\docType{package}
\title{
\packageTitle{scCNA}
}
\description{
\packageDescription{scCNA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scCNA}
\packageIndices{scCNA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{scCNA}

Maintainer: \packageMaintainer{scCNA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
